New commands, fixes, reorganization
Restructured OF warp commands to be under
`/ofw`. /ofw warp <warpName> to warp home
/ofw sethome <warpName> to set a group inspecific home
/ofw setwarp <group> <name> to set a group specific warp
/ofw list to list warps
