+++ src/openFactions/util/constants/MsgPrefix.java
@@ -10,5 +10,4 @@ public class MsgPrefix {
 	public final static String ERR = PLUGIN_NAME + ChatColor.DARK_RED + "["  + ChatColor.WHITE + "ERR" + ChatColor.DARK_RED +"]" + ChatColor.ITALIC +ChatColor.RED+  " ";
 	public final static String OK = PLUGIN_NAME + ChatColor.GREEN + "["  + ChatColor.WHITE + "OK" + ChatColor.GREEN +"]" + ChatColor.ITALIC + ChatColor.GREEN+ " ";
 	public final static String INFO = PLUGIN_NAME +ChatColor.AQUA+ "["  + ChatColor.WHITE + "INFO" + ChatColor.AQUA +"]"+ ChatColor.ITALIC + ChatColor.AQUA+ " ";
-	
 }
