+++ src/openFactions/util/constants/MsgSeparator.java
@@ -1,0 +1,13 @@
+package openFactions.util.constants;
+
+public class MsgSeparator {
+
+	/** : */
+	public final static String COLON_SINGLE = ":";
+	/** :: */
+	public final static String COLON_DOUBLE = "::";
+	/** ::: */
+	public final static String COLON_TRIPLE = ":::";
+	
+	public final static String SPACE_QUAD = "     ";
+}
